engine and dashboard resources
    - engine embeds resources in _embedded
    - the resources showed on the visualiser pages are clickable
    - double click on resource to show the resourceOwners in graph
    - various html and css changes and fixes
