+++ mindmaps-dashboard/static/index.html
@@ -8,7 +8,7 @@
     <link href='http://fonts.googleapis.com/css?family=Roboto:300,400,500,700,900' rel='stylesheet' type='text/css'>
 
     <!-- Page title -->
-    <title>MindmapsDB</title>
+    <title>GRAKN.AI</title>
 
     <!-- Vendor styles -->
     <link rel="stylesheet" href="vendor/fontawesome/css/font-awesome.css"/>
@@ -29,7 +29,7 @@
 <body>
 
 <!-- Wrapper-->
-<div id="app"></div>
+<div id="grakn-app"></div>
 <!-- End wrapper-->
 
 <!-- Vendor scripts -->
