+++ mindmaps-dashboard/static/styles/codemirror.css
@@ -242,7 +242,7 @@ div.CodeMirror span.CodeMirror-nonmatchingbracket {color: #f22;}
   border-width: 0;
   background: transparent;
   font-family: inherit;
-  font-size: inherit;
+  font-size: 15px;
   margin: 0;
   white-space: pre;
   word-wrap: normal;
